Функции-утилиты разбиты по выполнению
- функции для обработки данных
- функции для валидации форм
- функции-адаптеры
